requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
Werkzeug
psycopg2-binary
